M_PI needs to be enabled under Windows

+++ demo/iirdemo.cpp
@@ -8,7 +8,7 @@
 
 #include <stdio.h>
 
-int main (int,char**)
+int main(int, char**)
 {
 	const int order = 8;
 
@@ -20,18 +20,18 @@ int main (int,char**)
 	// Butterworth lowpass
 	Iir::Butterworth::LowPass<order> f;
 	double cutoff_frequency = 100; // Hz
-	f.setup (samplingrate, cutoff_frequency);
-	fimpulse = fopen("lp.dat","wt");
+	f.setup(samplingrate, cutoff_frequency);
+	fimpulse = fopen("lp.dat", "wt");
 	// let's simulated date streaming in
-	for(int i=0;i<1000;i++) 
+	for (int i = 0; i < 1000; i++)
 	{
-		double a=0;
-		if (i==10) a = 1;
+		double a = 0;
+		if (i == 10) a = 1;
 		double b = f.filter(a);
-		fprintf(fimpulse,"%e\n",b);
+		fprintf(fimpulse, "%e\n", b);
 	}
 	fclose(fimpulse);
-	
+
 	// RBJ highpass filter
 	// The Q factor determines the resonance at the
 	// cutoff frequency. If Q=1/sqrt(2) then the transition
@@ -40,29 +40,29 @@ int main (int,char**)
 	Iir::RBJ::HighPass hp;
 	const double hp_cutoff_frequency = 100;
 	const double hp_qfactor = 5;
-	hp.setup (samplingrate, hp_cutoff_frequency, hp_qfactor);
-	fimpulse = fopen("hp_rbj.dat","wt");
-	for(int i=0;i<1000;i++) 
+	hp.setup(samplingrate, hp_cutoff_frequency, hp_qfactor);
+	fimpulse = fopen("hp_rbj.dat", "wt");
+	for (int i = 0; i < 1000; i++)
 	{
-		double a=0;
-		if (i==10) a = 1;
+		double a = 0;
+		if (i == 10) a = 1;
 		double b = hp.filter(a);
-		fprintf(fimpulse,"%e\n",b);
+		fprintf(fimpulse, "%e\n", b);
 	}
 	fclose(fimpulse);
 
 	Iir::ChebyshevI::LowPass<8> lp_cheby1;
 	const float pass_ripple_db2 = 1; // dB
-	lp_cheby1.setup (samplingrate,
-			 cutoff_frequency,
-			 pass_ripple_db2);
-	fimpulse = fopen("lp_cheby1.dat","wt");
-	for(int i=0;i<1000;i++) 
+	lp_cheby1.setup(samplingrate,
+		cutoff_frequency,
+		pass_ripple_db2);
+	fimpulse = fopen("lp_cheby1.dat", "wt");
+	for (int i = 0; i < 1000; i++)
 	{
-		double a=0;
-		if (i==10) a = 1;
+		double a = 0;
+		if (i == 10) a = 1;
 		double b = lp_cheby1.filter(a);
-		fprintf(fimpulse,"%e\n",b);
+		fprintf(fimpulse, "%e\n", b);
 	}
 	fclose(fimpulse);
 
@@ -75,29 +75,29 @@ int main (int,char**)
 	// Setting cutoff to normalised f=0.1
 	double normalised_cutoff_freq = 0.1;
 	hp_cheby2.setupN(normalised_cutoff_freq,
-			 stop_ripple_dB);
-	fimpulse = fopen("hp_cheby2.dat","wt");
-	for(int i=0;i<1000;i++) 
+		stop_ripple_dB);
+	fimpulse = fopen("hp_cheby2.dat", "wt");
+	for (int i = 0; i < 1000; i++)
 	{
-		double a=0;
-		if (i==10) a = 1;
+		double a = 0;
+		if (i == 10) a = 1;
 		double b = hp_cheby2.filter(a);
-		fprintf(fimpulse,"%e\n",b);
+		fprintf(fimpulse, "%e\n", b);
 	}
 	fclose(fimpulse);
 
 	// Bandstop filter with default order
 	Iir::Butterworth::BandStop<> bsn;
-        const float norm_center_frequency = 0.1;
-        const float norm_frequency_width = 0.01;
-        bsn.setupN(norm_center_frequency, norm_frequency_width);
-	fimpulse = fopen("bs_norm.dat","wt");
-	for(int i=0;i<1000;i++) 
+	const double norm_center_frequency = 0.1;
+	const double norm_frequency_width = 0.01;
+	bsn.setupN(norm_center_frequency, norm_frequency_width);
+	fimpulse = fopen("bs_norm.dat", "wt");
+	for (int i = 0; i < 1000; i++)
 	{
-		double a=0;
-		if (i==10) a = 1;
+		double a = 0;
+		if (i == 10) a = 1;
 		double b = bsn.filter(a);
-		fprintf(fimpulse,"%e\n",b);
+		fprintf(fimpulse, "%e\n", b);
 	}
 	fclose(fimpulse);
 
@@ -118,17 +118,17 @@ int main (int,char**)
 	};
 	const int nSOS = sizeof(coeff) / sizeof(coeff[0]);
 	Iir::Custom::SOSCascade<nSOS> cust(coeff);
-	fimpulse = fopen("ellip.dat","wt");
-	for(int i=0;i<1000;i++) 
+	fimpulse = fopen("ellip.dat", "wt");
+	for (int i = 0; i < 1000; i++)
 	{
-		double a=0;
-		if (i==10) a = 1;
+		double a = 0;
+		if (i == 10) a = 1;
 		double b = cust.filter(a);
-		fprintf(fimpulse,"%e\n",b);
+		fprintf(fimpulse, "%e\n", b);
 	}
 	fclose(fimpulse);
 
-	
+
 
 	// generated with: "bessel_design.py"
 	const double coeff2[][6] = {
@@ -138,17 +138,17 @@ int main (int,char**)
 	};
 	const int nSOS2 = sizeof(coeff2) / sizeof(coeff2[0]);
 	Iir::Custom::SOSCascade<nSOS2> cust2(coeff2);
-	fimpulse = fopen("bessel.dat","wt");
-	for(int i=0;i<1000;i++) 
+	fimpulse = fopen("bessel.dat", "wt");
+	for (int i = 0; i < 1000; i++)
 	{
-		double a=0;
-		if (i==10) a = 1;
+		double a = 0;
+		if (i == 10) a = 1;
 		double b = cust2.filter(a);
-		fprintf(fimpulse,"%e\n",b);
+		fprintf(fimpulse, "%e\n", b);
 	}
 	fclose(fimpulse);
 
-	
+
 
 	printf("finished!\n");
 	printf("Now run `plot_impulse_fresponse.py` to display the impulse/frequency responses.\n");
